Module coroutine::coroutine

Basic single threaded Coroutine

use coroutine::{spawn, sched};

let coro = spawn(|| {
    println!("Before yield");

    // Yield back to its parent who resume this coroutine
    sched();

    println!("I am back!");
});

// Starts the Coroutine
coro.resume().ok().expect("Failed to resume");

println!("Back to main");

// Resume it
coro.resume().ok().expect("Failed to resume");

println!("Coroutine finished");

Structs

Builder
Coroutine configuration. Provides detailed control over the properties and behavior of new Coroutines.

Coroutine
A coroutine is nothing more than a (register context, stack) pair.

Handle
Handle of a Coroutine

Options
Coroutine spawn options

Enums

State
State of a Coroutine

Functions

current
Get the current Coroutine

resume
Resume a Coroutine

sched
Yield the current Coroutine

spawn
Spawn a new Coroutine

Type Definitions

ResumeResult
Return type of resuming.