Struct coroutine::coroutine::Builder

pub struct Builder {
    // some fields omitted
}

Coroutine configuration. Provides detailed control over the properties and behavior of new Coroutines.

let coro = Builder::new().name(format!("Coroutine #{}", 1))
                         .stack_size(4096)
                         .spawn(|| println!("Hello world!!"));

coro.resume().unwrap();

Methods

impl Builder

fn new() -> Builder

Generate the base configuration for spawning a Coroutine, from which configuration methods can be chained.

fn name(self, name: String) -> Builder

Name the Coroutine-to-be. Currently the name is used for identification only in panic messages.

fn stack_size(self, size: usize) -> Builder

Set the size of the stack for the new Coroutine.

fn spawn<F>(self, f: F) -> Handle where F: FnOnce() + Send + 'static

Spawn a new Coroutine, and return a handle for it.