Struct coroutine::coroutine_clonable::Coroutine

pub struct Coroutine {
    // some fields omitted
}

A coroutine is nothing more than a (register context, stack) pair.

Methods

impl Coroutine

fn spawn_opts<F>(f: F, opts: Options) -> Handle where F: FnOnce() + Send + 'static

Spawn a Coroutine with options

fn spawn<F>(f: F) -> Handle where F: FnOnce() + Send + 'static

Spawn a Coroutine with default options

fn yield_now(state: State)

Yield the current running Coroutine to its parent

fn sched()

Yield the current running Coroutine with Suspended state

fn block()

Yield the current running Coroutine with Blocked state

fn current() -> &'static Handle

Get a Handle to the current running Coroutine.

It is unsafe because it is an undefined behavior if you resume a Coroutine in more than one native thread.

fn name(&self) -> Option<&str>

Get the name of the Coroutine

fn panicking(&self) -> bool

Determines whether the current Coroutine is unwinding because of panic.

fn finished(&self) -> bool

Determines whether the Coroutine is finished