Enum coroutine::Error

pub enum Error {
    Finished,
    Waiting,
    Panicked,
    Panicking(Box<Any + Send>),
}

Resume Error

Variants

Finished

Coroutine is already finished

Waiting

Coroutine is waiting for its child to yield (state is Normal)

Panicked

Coroutine is panicked

Panicking

Coroutine is panicking, carry with the parameter of panic!()